--- a/src/DAP42.c
+++ b/src/DAP42.c
@@ -19,6 +19,7 @@
 #include <stdio.h>
 #include <string.h>
 
+#include <libopencm3/cm3/nvic.h>
 #include <libopencm3/stm32/desig.h>
 #include <libopencm3/stm32/iwdg.h>
 
@@ -71,6 +72,11 @@ static void on_dfu_request(void) {
     do_reset_to_dfu = true;
 }
 
+usbd_device* usbd_dev = NULL;
+void USB_IRQ_NAME(void) {
+    usbd_poll(usbd_dev);
+}
+
 int main(void) {
     if (DFU_AVAILABLE) {
         DFU_maybe_jump_to_bootloader();
@@ -105,7 +111,7 @@ int main(void) {
         cmp_set_usb_serial_number(serial);
     }
 
-    usbd_device* usbd_dev = cmp_usb_setup();
+    usbd_dev = cmp_usb_setup();
     DAP_app_setup(usbd_dev, &on_dfu_request);
 
     if (CDC_AVAILABLE) {
@@ -129,11 +135,14 @@ int main(void) {
 
     /* Enable the watchdog to enable DFU recovery from bad firmware images */
     iwdg_set_period_ms(1000);
-    iwdg_start();
+#warning "Disabling watchdog timer for now"
+    // iwdg_start();
+
+    /* Enable USB */
+    nvic_enable_irq(USB_NVIC_LINE);
 
     while (1) {
         iwdg_reset();
-        usbd_poll(usbd_dev);
 
         if (CDC_AVAILABLE) {
             cdc_uart_app_update();