… src/USB/bulk.c
… src/USB/bulk.h
… src/USB/cdc.c
… src/USB/composite_usb_conf.c
… src/USB/composite_usb_conf.h
… src/USB/hid.c
… src/USB/hid.h
… src/USB/usb21_standard.c
… src/USB/usb21_standard.h
… src/USB/winusb.c
… src/USB/winusb.h
… src/USB/winusb_defs.h
… src/stm32f042/brain3.3/config.h
… src/stm32f042/dap42/DAP/CMSIS_DAP_config.h
… src/stm32f042/dap42/config.h
… src/stm32f042/dap42dc/config.h
… src/stm32f042/dap42k6u/config.h
… src/stm32f042/kitchen42/config.h
… src/stm32f103/bluepill/config.h
… src/stm32f103/stlinkv2-1/config.h
… src/stm32f103/stlinkv2-dongle/config.h