Add support for bulk transfers, which are present in V2.

This greatly increases the transfer rate, with performance increases of 3-4x seen in light testing.

~~Unfortunately legacy HID transfers are slower as a result of this patch, for reasons which are currently unclear.~~ Edit: Fixed in a followup patch.