@xobs 2025-02-11
I managed to fix HID so that it works like it did before the rework.

For performance comparisons, here's the results of using `compare-sections` to check that a program matches what is on the target. This is better than measuring flash performance, since reads are usually only limited by the debugger speed.

Via Black Magic Debug Application:

Bulk:
```
bmd_crc32: 0x08000000+192 -> 4ms
bmd_crc32: 0x080000c0+52320 -> 701ms, 72 KiB/s
bmd_crc32: 0x0800cd20+15664 -> 204ms, 74 KiB/s
bmd_crc32: 0x08010a50+1028 -> 16ms, 62 KiB/s
```

HID:
```
bmd_crc32: 0x08000000+192 -> 14ms
bmd_crc32: 0x080000c0+52320 -> 2008ms, 25 KiB/s
bmd_crc32: 0x0800cd20+15664 -> 598ms, 25 KiB/s
bmd_crc32: 0x08010a50+1028 -> 44ms, 22 KiB/s
```

HID (before the patch):
```
bmd_crc32: 0x08000000+192 -> 13ms
bmd_crc32: 0x080000c0+52320 -> 2007ms, 25 KiB/s
bmd_crc32: 0x0800cd20+15664 -> 604ms, 25 KiB/s
bmd_crc32: 0x08010a50+1028 -> 43ms, 23 KiB/s
```