@devanlai 2025-02-11
Wow, those results are quite impressive. Thanks for taking the time to implement this.

I'll have to take some time to remember how the code worked (or aspired to work at least) and make sure everything still makes sense with your changes to the buffer management.

Aside from that, my main concern would be making sure that the compile-time resource constraint assertions are still up to date. In particular, it looks like the new bulk endpoints are not represented in the packet memory area calculations in `composite_usb_conf.c`.