@devanlai 2025-02-14
Alright, I think I've reloaded some of the context around the USB handling now.

In no particular order, my general thoughts:

- I like how you've slimmed it down to one circular buffer for the incoming DAP commands. It was definitely unnecessary to have dedicated slots for both the request and the response, given that they only briefly both exist simultaneously.
- I can believe that under normal testing, sending the response over the bulk interface immediately will not block because the host is likely to poll/drain the bulk interface faster than we can execute DAP commands and output the responses.
- I would feel better if it always put the response into the circular buffer to handle the edge case where requests come in faster than the responses are drained.
- In the unlikely event that requests come in approximately as fast as they're processed, it might be possible to sit in the command processing loop in `DAP_app_update()` indefinitely, starving the rest of the superloop and triggering the watchdog. I'd be curious if there's a performance hit to capping the number of commands processed to per invocation to at most one or two to guard against that.
- I forget whether zero-length packets are actually required for CMSIS-DAP. They're definitely needed for CDC-ACM which can carry arbitrary unframed data, but it may be unnecessary for CMSIS-DAP where it can be understood at the protocol level that each response gets its own standalone packet.
- Not related to any of your changes, but I recall now that there is a race condition between trying to fill the HID IN endpoint buffer from `DAP_app_update()` vs the USB callback for the HID IN endpoint trying to refill it from the buffer after the endpoint is polled. I see that your scheme for the bulk endpoint of only populating the endpoint from from `DAP_app_update()` avoids that.

I'll have to think about it some more and dust off my dev boards to do some experiments.