@xobs 2025-02-14
Regarding the ZLP requirement -- it definitely was required when I did it. If you send 64 bytes on a USB FS link, then the host must ask for the rest of the buffer. You indicate that there is no more data by sending a ZLP. I didn't get any data before I added that.

I believe it's the case for Interrupt endpoints that they can only transmit at most 64 bytes, so there's no need for a ZLP there.

Perhaps you're right in that the clients I've worked with always do an OUT followed immediately by an IN, and I can definitely see the case where they pipeline multiple OUT requests, especially for very long sequences. I'll look into pipelining bulk requests. I left that bomb in there if the IN request fails, and that does make me a bit uncomfortable.

There's also the separate issue in that with the bulk endpoints, we probably should add the Microsoft descriptors. I'll work on that after I get pipelined bulk transfers working.