@devanlai 2025-02-14
> Regarding the ZLP requirement -- it definitely was required when I did it. If you send 64 bytes on a USB FS link, then the host must ask for the rest of the buffer. You indicate that there is no more data by sending a ZLP. I didn't get any data before I added that.
> 
> I believe it's the case for Interrupt endpoints that they can only transmit at most 64 bytes, so there's no need for a ZLP there.

I think you're right here - now that I think about it more, it's up to the host to specify how many bytes they're expecting up front, and if they don't always know in advance, they have to rely on ZLPs to delimit the end. It might be possible to infer the response size from the request, but it's probably safest for them to rely on ZLPs for maximum compatibility.

> There's also the separate issue in that with the bulk endpoints, we probably should add the Microsoft descriptors. I'll work on that after I get pipelined bulk transfers working.

I have an [old branch](https://github.com/devanlai/dap42/tree/winusb-wip) that was adding the Microsoft descriptors for the DFU interface together with some webusb stuff. It should be possible to extract just the USB 2.1 BOS and MS OS 2.0 descriptor functionality and adapt the descriptors for just CMSIS-DAP 2.0 instead. (And as long as the webusb landing URL functionality is excluded, we won't go over the hardcoded limit of at most 4 control callbacks).