@xobs 2025-02-14
I'm not right. ZLP was completely broken. I've removed it, and added support for pipelining.

I put together this completely awful gist that helped me to trigger various cases. It works now: https://gist.github.com/xobs/1bc85cbe7da8aef31c3b5443d752f2cf

I can have a look at that patch for winusb.