@xobs 2025-02-14
Actually, I think winusb should go in a separate PR. I'm happy with this for now, given it passes the tests I've thrown at it.

I'll work on winusb separately. Was there any reason you didn't merge that branch? Do you recall what issues you ran into? I realize that was nine years ago!