@xobs 2025-02-14
I see. Then with the addition of bulk endpoints, it becomes more important to allow Windows to take advantage of the new interface without having to resort to Zadig, whereas that wasn't an issue before.