@devanlai 2025-02-16
It looks like I'll have to do some more testing and debugging on my side:
1. Unrelated to your changes (even on the mainline branch), I'm seeing issues with the HID CMSIS-DAP interface timing out on Windows, so I'll have to get my proper development environment setup again.
2. My STLink/v2 clone dongles aren't enumerating correctly with the bulk interface enabled, so I'll have to dig into that.

I've also started a branch at [cmsis-dap-bulk](https://github.com/devanlai/dap42/tree/cmsis-dap-bulk) that aligns things a bit more closely with how I had originally planned to manage the USB interrupt handling. It remains to be seen whether this is a bad strategy with hidden race conditions.