@xobs 2025-02-16
1. Strange. I haven't checked Windows at all so I haven't seen it. Do you have suggestions as to how to trigger those conditions?
2. I have o STM32F103 harware. Do you have recommendations on boards to get?

I'm working to gain access to a USB Beagle. I really should have bought my own by now, what with how useful they are.

I like the changes to buffers, assuming they work. The one thing I'd change is to mark `process_head` as `volatile` since it's updated both from an interrupt and a non-interrupt context. Otherwise the optimizer has a tendency to get clever.